Move swaggerdoc generator
